int's veranderd naar double's, nog steeds verkeerde resultaat

+++ src/com/codevillain679/behavioral/strategy/CalculateStrategy.java
@@ -1,5 +1,5 @@
 package com.codevillain679.behavioral.strategy;
 
 public interface CalculateStrategy {
-    public int Calculate(int a, int b);
+    public double Calculate(double a, double b);
 }
