+++ src/com/codevillain679/behavioral/strategy/Subtract.java
@@ -3,9 +3,9 @@ package com.codevillain679.behavioral.strategy;
 public class Subtract implements CalculateStrategy {
 
     @Override
-    public int Calculate(int a, int b) {
-        int x;
-        x = a-b;
+    public double Calculate(double a, double b) {
+        double x;
+        x = a - b;
         return x;
     }
 }
